pyproject.toml
[tool.poetry]
name = "class_management_back"
version = "0.1.0"
description = ""
authors = ["Vitor <vitor.ricoy@dcc.ufmg.br>"]
packages = [
    {include = "class_management_back", from = "src"}
]

[tool.poetry.dependencies]
python = "^3.10"
Flask = {version = "^2.0.2", extras = ["async"] }
Flask-HTTPAuth = "^4.5.0"
passlib = "^1.7.4"
SQLAlchemy = "^1.3"
psycopg2 = "^2.9.3"
flask-restx = "^0.5.1"
Flask-RESTful = "^0.3.9"
Flask-Cors = "^3.0.10"
pydantic = {extras = ["email"], version = "^1.10.5"}
PyJWT = "^2.6.0"
pandas = "^2.0.1"
numpy = "^1.24.3"
joblib = "^1.2.0"
scikit-learn = "^1.2.2"

[tool.pyright]
include = ["src"]
exclude = ["**/__pycache__"]

pythonVersion = "3.10"
pythonPlatform = "Linux"

reportMissingImports = false
reportInvalidStringEscapeSequence = false

[tool.black]
line-length = 79

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
